기존 pull 받은 파일 commit

+++ src/main/java/bucheon/leafy/application/controller/FeedController.java
@@ -1,15 +1,20 @@
 package bucheon.leafy.application.controller;
 
 import bucheon.leafy.application.service.FeedService;
+import bucheon.leafy.config.AuthUser;
 import bucheon.leafy.domain.feed.request.FeedRequest;
 import bucheon.leafy.domain.feed.response.FeedResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+
 
 @Tag(name = "피드")
 @RestController
@@ -40,19 +45,35 @@ public class FeedController {
 
     @Operation(summary = "피드 등록")
     @PostMapping
-    public ResponseEntity<Long> saveFeed(@RequestBody FeedRequest request) {
-        return ResponseEntity.ok().body(service.saveFeed(request));
+    @PreAuthorize("hasAnyRole('MEMBER', 'ADMIN')")
+    public ResponseEntity<Long> saveFeed(@AuthenticationPrincipal AuthUser user, @RequestBody FeedRequest request) {
+        Long userId = user.getUserId();
+        return ResponseEntity.ok().body(service.saveFeed(userId, request));
     }
 
     @Operation(summary = "피드 수정")
     @PutMapping("/{feedId}")
-    public ResponseEntity<Long> updateFeed(@PathVariable Long feedId, @RequestBody FeedRequest request) {
-        return ResponseEntity.ok().body(service.updateFeed(feedId, request));
+    @PreAuthorize("hasAnyRole('MEMBER', 'ADMIN')")
+    public ResponseEntity<Long> updateFeed(@AuthenticationPrincipal AuthUser user, @PathVariable Long feedId, @RequestBody FeedRequest request) {
+        Long userId = user.getUserId();
+        FeedResponse response = service.getFeedById(feedId);
+        if( userId.equals(response.getUserId()) ) {
+            return ResponseEntity.ok().body(service.updateFeed(userId, feedId, request));
+        } else {
+            throw new AccessDeniedException("수정 권한이 없습니다.");
+        }
     }
 
     @Operation(summary = "피드 삭제")
     @DeleteMapping("/{feedId}")
-    public ResponseEntity<Boolean> deleteFeed(@PathVariable Long feedId) {
-        return ResponseEntity.ok().body(service.deleteFeed(feedId));
+    @PreAuthorize("hasAnyRole('MEMBER', 'ADMIN')")
+    public ResponseEntity<Boolean> deleteFeed(@AuthenticationPrincipal AuthUser user, @PathVariable Long feedId) {
+        Long userId = user.getUserId();
+        FeedResponse response = service.getFeedById(feedId);
+        if( userId.equals(response.getUserId()) ) {
+            return ResponseEntity.ok().body(service.deleteFeed(feedId));
+        } else {
+            throw new AccessDeniedException("삭제 권한이 없습니다.");
+        }
     }
 }
